Fix pm3 2 07 + 10

+++ contracts/IPredictionMarketV3_2.sol
@@ -16,7 +16,7 @@ interface IPredictionMarketV3_2 {
 
   struct CreateMarketDescription {
     uint256 value;
-    uint256 closesAt;
+    uint32 closesAt;
     uint256 outcomes;
     address token;
     uint256[] distribution;
@@ -27,7 +27,7 @@ interface IPredictionMarketV3_2 {
     Fees sellFees;
     address treasury;
     address distributor;
-    uint256 realitioTimeout;
+    uint32 realitioTimeout;
     address manager;
   }
